+++ presidio-image-redactor/app.py
@@ -60,7 +60,7 @@ class Server:
 
             elif request.files and "image" in request.files:
                 im = Image.open(request.files.get("image"))
-                redacted_image = self.engine.redact(im, color_fill)
+                redacted_image = self.engine.redact(im, color_fill, score_threshold=0.4)
                 img_byte_arr = image_to_byte_array(redacted_image, im.format)
                 return Response(img_byte_arr,
                                 mimetype="application/octet-stream")